install ffmpeg to docker

+++ src/main.ts
@@ -49,32 +49,28 @@ async function loadExternalPlugins() {
 function commandCatcher(lastMessage: WAMessage) {
   loadedCommands.map(async (command) => {
     let message: string = '';
-    if (lastMessage.message?.conversation) {
-      // If user simply just use the command
-      message = lastMessage.message.conversation;
-    }
 
-    if (lastMessage.message?.extendedTextMessage) {
-      // If user quoted a message
-      if (lastMessage.message.extendedTextMessage.text) {
-        message = lastMessage.message.extendedTextMessage.text;
+    if (lastMessage.key.fromMe) {
+      if (lastMessage.message?.conversation) {
+        // If user simply just use the command
+        message = lastMessage.message.conversation;
       }
-    }
 
-    if (lastMessage.message?.imageMessage?.caption) {
-      // If replying a image message with command
-      message = lastMessage?.message?.imageMessage.caption;
-    }
-    console.log(command.pattern);
+      if (lastMessage.message?.imageMessage?.caption) {
+        // If replying a image message with command
+        message = lastMessage?.message?.imageMessage.caption;
+      }
+      console.log(command.pattern);
 
-    // Check if message contains any command with help of regex.
-    let match: RegExpMatchArray | null = message.match(command.pattern);
-    // console.log(match);
-    if (match) {
-      console.log(match);
-      const client = new Message(bot, lastMessage);
-      await client.delete();
-      command.function(client, match);
+      // Check if message contains any command with help of regex.
+      let match: RegExpMatchArray | null = message.match(command.pattern);
+      // console.log(match);
+      if (match) {
+        console.log(match);
+        const client = new Message(bot, lastMessage);
+        await client.delete();
+        command.function(client, match);
+      }
     }
   });
 }
